Refactor raft node to break it away from worker into the new conn package.
