+++ cmd/dgraph/main.go
@@ -46,6 +46,7 @@ import (
 	"google.golang.org/grpc"
 
 	"github.com/cockroachdb/cmux"
+	"github.com/dgraph-io/dgraph/conn"
 	"github.com/dgraph-io/dgraph/dgraph"
 	"github.com/dgraph-io/dgraph/gql"
 	"github.com/dgraph-io/dgraph/group"
@@ -651,6 +652,7 @@ func serveGRPC(l net.Listener) {
 		grpc.MaxSendMsgSize(x.GrpcMaxSize),
 		grpc.MaxConcurrentStreams(1000))
 	protos.RegisterDgraphServer(s, &dgraph.Server{})
+	protos.RegisterRaftServer(s, &conn.RaftServer{})
 	err := s.Serve(l)
 	log.Printf("gRpc server stopped : %s", err.Error())
 	s.GracefulStop()
