+++ worker/draft.go
@@ -18,7 +18,6 @@
 package worker
 
 import (
-	"bytes"
 	"encoding/binary"
 	"fmt"
 	"log"
@@ -42,65 +41,6 @@ import (
 const (
 	errorNodeIDExists = "Error Node ID already exists in the cluster"
 )
-
-type peerPoolEntry struct {
-	// Never the empty string.  Possibly a bogus address -- bad port number, the value
-	// of *myAddr, or some screwed up Raft config.
-	addr string
-	// An owning reference to a pool for this peer (or nil if addr is sufficiently bogus).
-	poolOrNil *conn.Pool
-}
-
-// peerPool stores the peers' addresses and our connections to them.  It has exactly one
-// entry for every peer other than ourselves.  Some of these peers might be unreachable or
-// have bogus (but never empty) addresses.
-type peerPool struct {
-	sync.RWMutex
-	peers map[uint64]peerPoolEntry
-}
-
-var (
-	errNoPeerPoolEntry = fmt.Errorf("no peerPool entry")
-	errNoPeerPool      = fmt.Errorf("no peerPool pool, could not connect")
-)
-
-// getPool returns the non-nil pool for a peer.  This might error even if get(id)
-// succeeds, if the pool is nil.  This happens if the peer was configured so badly (it had
-// a totally bogus addr) we can't make a pool.  (A reasonable refactoring would have us
-// make a pool, one that has a nil gRPC connection.)
-//
-// You must call pools().release on the pool.
-func (p *peerPool) getPool(id uint64) (*conn.Pool, error) {
-	p.RLock()
-	defer p.RUnlock()
-	ent, ok := p.peers[id]
-	if !ok {
-		return nil, errNoPeerPoolEntry
-	}
-	if ent.poolOrNil == nil {
-		return nil, errNoPeerPool
-	}
-	ent.poolOrNil.AddOwner()
-	return ent.poolOrNil, nil
-}
-
-func (p *peerPool) get(id uint64) (string, bool) {
-	p.RLock()
-	defer p.RUnlock()
-	ret, ok := p.peers[id]
-	return ret.addr, ok
-}
-
-func (p *peerPool) set(id uint64, addr string, pl *conn.Pool) {
-	p.Lock()
-	defer p.Unlock()
-	if old, ok := p.peers[id]; ok {
-		if old.poolOrNil != nil {
-			conn.Get().Release(old.poolOrNil)
-		}
-	}
-	p.peers[id] = peerPoolEntry{addr, pl}
-}
 
 type proposalCtx struct {
 	ch  chan error
@@ -178,32 +118,16 @@ func (p *proposals) Has(pid uint32) bool {
 	return has
 }
 
-type sendmsg struct {
-	to   uint64
-	data []byte
-}
-
 type node struct {
-	x.SafeMutex
-
-	// SafeMutex is for fields which can be changed after init.
-	_confState *raftpb.ConfState
-	_raft      raft.Node
+	*conn.Node
 
 	// Fields which are never changed after init.
-	cfg         *raft.Config
-	applyCh     chan raftpb.Entry
-	ctx         context.Context
-	stop        chan struct{} // to send the stop signal to Run
-	done        chan struct{} // to check whether node is running or not
-	gid         uint32
-	id          uint64
-	messages    chan sendmsg
-	peers       peerPool
-	props       proposals
-	raftContext *protos.RaftContext
-	store       *raft.MemoryStorage
-	wal         *raftwal.Wal
+	applyCh chan raftpb.Entry
+	ctx     context.Context
+	stop    chan struct{} // to send the stop signal to Run
+	done    chan struct{} // to check whether node is running or not
+	gid     uint32
+	props   proposals
 
 	canCampaign bool
 	// applied is used to keep track of the applied RAFT proposals.
@@ -213,142 +137,34 @@ type node struct {
 	sch     *scheduler
 }
 
-// SetRaft would set the provided raft.Node to this node.
-// It would check fail if the node is already set.
-func (n *node) SetRaft(r raft.Node) {
-	n.Lock()
-	defer n.Unlock()
-	x.AssertTrue(n._raft == nil)
-	n._raft = r
-}
-
-// Raft would return back the raft.Node stored in the node.
-func (n *node) Raft() raft.Node {
-	n.RLock()
-	defer n.RUnlock()
-	return n._raft
-}
-
-// SetConfState would store the latest ConfState generated by ApplyConfChange.
-func (n *node) SetConfState(cs *raftpb.ConfState) {
-	n.Lock()
-	defer n.Unlock()
-	n._confState = cs
-}
-
-// ConfState would return the latest ConfState stored in node.
-func (n *node) ConfState() *raftpb.ConfState {
-	n.RLock()
-	defer n.RUnlock()
-	return n._confState
-}
-
 func newNode(gid uint32, id uint64, myAddr string) *node {
 	x.Printf("Node with GroupID: %v, ID: %v\n", gid, id)
 
-	peers := peerPool{
-		peers: make(map[uint64]peerPoolEntry),
-	}
-	props := proposals{
-		ids: make(map[uint32]*proposalCtx),
-	}
-
-	store := raft.NewMemoryStorage()
 	rc := &protos.RaftContext{
 		Addr:  myAddr,
 		Group: gid,
 		Id:    id,
 	}
 
-	n := &node{
-		ctx:   context.Background(),
-		id:    id,
-		gid:   gid,
-		store: store,
-		cfg: &raft.Config{
-			ID:              id,
-			ElectionTick:    100, // 200 ms if we call Tick() every 20 ms.
-			HeartbeatTick:   1,   // 20 ms if we call Tick() every 20 ms.
-			Storage:         store,
-			MaxSizePerMsg:   256 << 10,
-			MaxInflightMsgs: 256,
-			Logger:          &raft.DefaultLogger{Logger: x.Logger},
-		},
-		// processConfChange etc are not throttled so some extra delta, so that we don't
-		// block tick when applyCh is full
-		applyCh:     make(chan raftpb.Entry, Config.NumPendingProposals+1000),
-		peers:       peers,
-		props:       props,
-		raftContext: rc,
-		messages:    make(chan sendmsg, 1000),
-		stop:        make(chan struct{}),
-		done:        make(chan struct{}),
+	m := conn.NewNode(rc)
+	n := &node{Node: m}
+
+	n.gid = gid
+	// processConfChange etc are not throttled so some extra delta, so that we don't
+	// block tick when applyCh is full
+	n.applyCh = make(chan raftpb.Entry, Config.NumPendingProposals+1000)
+	n.props = proposals{
+		ids: make(map[uint32]*proposalCtx),
 	}
+	n.stop = make(chan struct{})
+	n.done = make(chan struct{})
+
 	n.sch = new(scheduler)
 	n.sch.init(n)
 	n.applied = x.WaterMark{Name: fmt.Sprintf("Committed: Group %d", n.gid)}
 	n.applied.Init()
 
 	return n
-}
-
-// Never returns ("", true)
-func (n *node) GetPeer(pid uint64) (string, bool) {
-	return n.peers.get(pid)
-}
-
-// You must call release on the pool.  Can error for some pid's for which GetPeer
-// succeeds.
-func (n *node) GetPeerPool(pid uint64) (*conn.Pool, error) {
-	return n.peers.getPool(pid)
-}
-
-// addr must not be empty.
-func (n *node) SetPeer(pid uint64, addr string, poolOrNil *conn.Pool) {
-	x.AssertTruef(addr != "", "SetPeer for peer %d has empty addr.", pid)
-	n.peers.set(pid, addr, poolOrNil)
-}
-
-// Connects the node and makes its peerPool refer to the constructed pool and address
-// (possibly updating ourselves from the old address.)  (Unless pid is ourselves, in which
-// case this does nothing.)
-func (n *node) Connect(pid uint64, addr string) {
-	if pid == n.id {
-		return
-	}
-	if paddr, ok := n.GetPeer(pid); ok && paddr == addr {
-		// Already connected.
-		return
-	}
-	// Here's what we do.  Right now peerPool maps peer node id's to addr values.  If
-	// a *pool can be created, good, but if not, we still create a peerPoolEntry with
-	// a nil *pool.
-	if addr == Config.MyAddr {
-		// TODO: Note this fact in more general peer health info somehow.
-		x.Printf("Peer %d claims same host as me\n", pid)
-		n.SetPeer(pid, addr, nil)
-		return
-	}
-	p := conn.Get().Connect(addr)
-	n.SetPeer(pid, addr, p)
-}
-
-func (n *node) AddToCluster(ctx context.Context, pid uint64) error {
-	addr, ok := n.GetPeer(pid)
-	x.AssertTruef(ok, "Unable to find conn pool for peer: %d", pid)
-	rc := &protos.RaftContext{
-		Addr:  addr,
-		Group: n.raftContext.Group,
-		Id:    pid,
-	}
-	rcBytes, err := rc.Marshal()
-	x.Check(err)
-	return n.Raft().ProposeConfChange(ctx, raftpb.ConfChange{
-		ID:      pid,
-		Type:    raftpb.ConfChangeAddNode,
-		NodeID:  pid,
-		Context: rcBytes,
-	})
 }
 
 type header struct {
@@ -448,103 +264,6 @@ func (n *node) ProposeAndWait(ctx context.Context, proposal *protos.Proposal) er
 		}
 	}
 	return err
-}
-
-func (n *node) send(m raftpb.Message) {
-	x.AssertTruef(n.id != m.To, "Seding message to itself")
-	data, err := m.Marshal()
-	x.Check(err)
-	if m.Type != raftpb.MsgHeartbeat && m.Type != raftpb.MsgHeartbeatResp {
-		x.Printf("\t\tSENDING: %v %v-->%v\n", m.Type, m.From, m.To)
-	}
-	select {
-	case n.messages <- sendmsg{to: m.To, data: data}:
-		// pass
-	default:
-		// TODO: It's bad to fail like this.
-		x.Fatalf("Unable to push messages to channel in send")
-	}
-}
-
-const (
-	messageBatchSoftLimit = 10000000
-)
-
-func (n *node) batchAndSendMessages() {
-	batches := make(map[uint64]*bytes.Buffer)
-	for {
-		totalSize := 0
-		sm := <-n.messages
-	slurp_loop:
-		for {
-			var buf *bytes.Buffer
-			if b, ok := batches[sm.to]; !ok {
-				buf = new(bytes.Buffer)
-				batches[sm.to] = buf
-			} else {
-				buf = b
-			}
-			totalSize += 4 + len(sm.data)
-			x.Check(binary.Write(buf, binary.LittleEndian, uint32(len(sm.data))))
-			x.Check2(buf.Write(sm.data))
-
-			if totalSize > messageBatchSoftLimit {
-				// We limit the batch size, but we aren't pushing back on
-				// n.messages, because the loop below spawns a goroutine
-				// to do its dirty work.  This is good because right now
-				// (*node).send fails(!) if the channel is full.
-				break
-			}
-
-			select {
-			case sm = <-n.messages:
-			default:
-				break slurp_loop
-			}
-		}
-
-		for to, buf := range batches {
-			if buf.Len() == 0 {
-				continue
-			}
-			data := make([]byte, buf.Len())
-			copy(data, buf.Bytes())
-			go n.doSendMessage(to, data)
-			buf.Reset()
-		}
-	}
-}
-
-func (n *node) doSendMessage(to uint64, data []byte) {
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	pool, err := n.GetPeerPool(to)
-	if err != nil {
-		// No such peer exists or we got handed a bogus config (bad addr), so we
-		// can't send messages to this peer.
-		return
-	}
-	defer conn.Get().Release(pool)
-	conn := pool.Get()
-
-	c := protos.NewWorkerClient(conn)
-	p := &protos.Payload{Data: data}
-
-	ch := make(chan error, 1)
-	go func() {
-		_, err = c.RaftMessage(ctx, p)
-		ch <- err
-	}()
-
-	select {
-	case <-ctx.Done():
-		return
-	case <-ch:
-		// We don't need to do anything if we receive any error while sending message.
-		// RAFT would automatically retry.
-		return
-	}
 }
 
 func (n *node) processMutation(pid uint32, index uint64, edge *protos.DirectedEdge) error {
@@ -652,28 +371,6 @@ func (n *node) processApplyCh() {
 	}
 }
 
-func (n *node) saveToStorage(s raftpb.Snapshot, h raftpb.HardState,
-	es []raftpb.Entry) {
-	if !raft.IsEmptySnap(s) {
-		le, err := n.store.LastIndex()
-		if err != nil {
-			log.Fatalf("While retrieving last index: %v\n", err)
-		}
-		if s.Metadata.Index <= le {
-			return
-		}
-
-		if err := n.store.ApplySnapshot(s); err != nil {
-			log.Fatalf("Applying snapshot: %v", err)
-		}
-	}
-
-	if !raft.IsEmptyHardState(h) {
-		n.store.SetHardState(h)
-	}
-	n.store.Append(es)
-}
-
 func (n *node) retrieveSnapshot(peerID uint64) {
 	pool, err := n.GetPeerPool(peerID)
 	if err != nil {
@@ -684,7 +381,7 @@ func (n *node) retrieveSnapshot(peerID uint64) {
 	defer conn.Get().Release(pool)
 
 	// Get index of last committed.
-	lastIndex, err := n.store.LastIndex()
+	lastIndex, err := n.Store.LastIndex()
 	x.Checkf(err, "Error while getting last index")
 	// Wait for watermarks to sync since populateShard writes directly to db, otherwise
 	// the values might get overwritten
@@ -710,7 +407,7 @@ func (n *node) Run() {
 	// See also our configuration of HeartbeatTick and ElectionTick.
 	ticker := time.NewTicker(20 * time.Millisecond)
 	defer ticker.Stop()
-	rcBytes, err := n.raftContext.Marshal()
+	rcBytes, err := n.RaftContext.Marshal()
 	x.Check(err)
 	for {
 		select {
@@ -733,16 +430,16 @@ func (n *node) Run() {
 				for _, msg := range rd.Messages {
 					// NOTE: We can do some optimizations here to drop messages.
 					msg.Context = rcBytes
-					n.send(msg)
+					n.Send(msg)
 				}
 			}
 
 			// First store the entries, then the hardstate and snapshot.
-			x.Check(n.wal.Store(n.gid, rd.HardState, rd.Entries))
-			x.Check(n.wal.StoreSnapshot(n.gid, rd.Snapshot))
+			x.Check(n.Wal.Store(n.gid, rd.HardState, rd.Entries))
+			x.Check(n.Wal.StoreSnapshot(n.gid, rd.Snapshot))
 
 			// Now store them in the in-memory store.
-			n.saveToStorage(rd.Snapshot, rd.HardState, rd.Entries)
+			n.SaveToStorage(rd.Snapshot, rd.HardState, rd.Entries)
 
 			if !raft.IsEmptySnap(rd.Snapshot) {
 				// We don't send snapshots to other nodes. But, if we get one, that means
@@ -751,9 +448,9 @@ func (n *node) Run() {
 				var rc protos.RaftContext
 				x.Check(rc.Unmarshal(rd.Snapshot.Data))
 				x.AssertTrue(rc.Group == n.gid)
-				if rc.Id != n.id {
+				if rc.Id != n.Id {
 					// NOTE: Retrieving snapshot here is OK, after storing it above in WAL, because
-					// rc.Id != n.id.
+					// rc.Id != n.Id.
 					x.Printf("-------> SNAPSHOT [%d] from %d\n", n.gid, rc.Id)
 					// It's ok to block tick while retrieving snapshot, since it's a follower
 					n.retrieveSnapshot(rc.Id)
@@ -795,7 +492,7 @@ func (n *node) Run() {
 				for _, msg := range rd.Messages {
 					// NOTE: We can do some optimizations here to drop messages.
 					msg.Context = rcBytes
-					n.send(msg)
+					n.Send(msg)
 				}
 			}
 			n.Raft().Advance()
@@ -841,10 +538,6 @@ func (n *node) Stop() {
 	<-n.done // wait for Run to respond.
 }
 
-func (n *node) Step(ctx context.Context, msg raftpb.Message) error {
-	return n.Raft().Step(ctx, msg)
-}
-
 func (n *node) snapshotPeriodically() {
 	if n.gid == 0 {
 		// Group zero is dedicated for membership information, whose state we don't persist.
@@ -877,7 +570,7 @@ func (n *node) snapshot(skip uint64) {
 	water := posting.SyncMarkFor(n.gid)
 	le := water.DoneUntil()
 
-	existing, err := n.store.Snapshot()
+	existing, err := n.Store.Snapshot()
 	x.Checkf(err, "Unable to get existing snapshot")
 
 	si := existing.Metadata.Index
@@ -888,13 +581,13 @@ func (n *node) snapshot(skip uint64) {
 	if tr, ok := trace.FromContext(n.ctx); ok {
 		tr.LazyPrintf("Taking snapshot for group: %d at watermark: %d\n", n.gid, snapshotIdx)
 	}
-	rc, err := n.raftContext.Marshal()
+	rc, err := n.RaftContext.Marshal()
 	x.Check(err)
 
-	s, err := n.store.CreateSnapshot(snapshotIdx, n.ConfState(), rc)
+	s, err := n.Store.CreateSnapshot(snapshotIdx, n.ConfState(), rc)
 	x.Checkf(err, "While creating snapshot")
-	x.Checkf(n.store.Compact(snapshotIdx), "While compacting snapshot")
-	x.Check(n.wal.StoreSnapshot(n.gid, s))
+	x.Checkf(n.Store.Compact(snapshotIdx), "While compacting snapshot")
+	x.Check(n.Wal.StoreSnapshot(n.gid, s))
 }
 
 func (n *node) joinPeers() {
@@ -917,15 +610,15 @@ func (n *node) joinPeers() {
 
 	conn := pool.Get()
 
-	c := protos.NewWorkerClient(conn)
+	c := protos.NewRaftClient(conn)
 	x.Printf("Calling JoinCluster")
-	_, err = c.JoinCluster(n.ctx, n.raftContext)
+	_, err = c.JoinCluster(n.ctx, n.RaftContext)
 	x.Checkf(err, "Error while joining cluster")
 	x.Printf("Done with JoinCluster call\n")
 }
 
 func (n *node) initFromWal(wal *raftwal.Wal) (restart bool, rerr error) {
-	n.wal = wal
+	n.Wal = wal
 
 	var sp raftpb.Snapshot
 	sp, rerr = wal.Snapshot(n.gid)
@@ -936,7 +629,7 @@ func (n *node) initFromWal(wal *raftwal.Wal) (restart bool, rerr error) {
 	if !raft.IsEmptySnap(sp) {
 		x.Printf("Found Snapshot: %+v\n", sp)
 		restart = true
-		if rerr = n.store.ApplySnapshot(sp); rerr != nil {
+		if rerr = n.Store.ApplySnapshot(sp); rerr != nil {
 			return
 		}
 		term = sp.Metadata.Term
@@ -953,7 +646,7 @@ func (n *node) initFromWal(wal *raftwal.Wal) (restart bool, rerr error) {
 	if !raft.IsEmptyHardState(hd) {
 		x.Printf("Found hardstate: %+v\n", hd)
 		restart = true
-		if rerr = n.store.SetHardState(hd); rerr != nil {
+		if rerr = n.Store.SetHardState(hd); rerr != nil {
 			return
 		}
 	}
@@ -967,7 +660,7 @@ func (n *node) initFromWal(wal *raftwal.Wal) (restart bool, rerr error) {
 	if len(es) > 0 {
 		restart = true
 	}
-	rerr = n.store.Append(es)
+	rerr = n.Store.Append(es)
 	return
 }
 
@@ -982,17 +675,17 @@ func (n *node) InitAndStartNode(wal *raftwal.Wal) {
 		if !found && groups().HasPeer(n.gid) {
 			n.joinPeers()
 		}
-		n.SetRaft(raft.RestartNode(n.cfg))
+		n.SetRaft(raft.RestartNode(n.Cfg))
 
 	} else {
 		x.Printf("New Node for group: %d\n", n.gid)
 		if groups().HasPeer(n.gid) {
 			n.joinPeers()
-			n.SetRaft(raft.StartNode(n.cfg, nil))
+			n.SetRaft(raft.StartNode(n.Cfg, nil))
 
 		} else {
-			peers := []raft.Peer{{ID: n.id}}
-			n.SetRaft(raft.StartNode(n.cfg, peers))
+			peers := []raft.Peer{{ID: n.Id}}
+			n.SetRaft(raft.StartNode(n.Cfg, peers))
 			// Trigger election, so this node can become the leader of this single-node cluster.
 			n.canCampaign = true
 		}
@@ -1002,7 +695,7 @@ func (n *node) InitAndStartNode(wal *raftwal.Wal) {
 	// TODO: Find a better way to snapshot, so we don't lose the membership
 	// state information, which isn't persisted.
 	go n.snapshotPeriodically()
-	go n.batchAndSendMessages()
+	go n.BatchAndSendMessages()
 }
 
 func (n *node) AmLeader() bool {
@@ -1011,89 +704,4 @@ func (n *node) AmLeader() bool {
 	}
 	r := n.Raft()
 	return r.Status().Lead == r.Status().ID
-}
-
-var (
-	errNoNode = fmt.Errorf("No node has been set up yet")
-)
-
-func applyMessage(ctx context.Context, msg raftpb.Message) error {
-	var rc protos.RaftContext
-	x.Check(rc.Unmarshal(msg.Context))
-	node := groups().Node(rc.Group)
-	if node == nil {
-		// Maybe we went down, went back up, reconnected, and got an RPC
-		// message before we set up Raft?
-		return errNoNode
-	}
-	node.Connect(msg.From, rc.Addr)
-
-	c := make(chan error, 1)
-	go func() { c <- node.Step(ctx, msg) }()
-
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case err := <-c:
-		return err
-	}
-}
-
-func (w *grpcWorker) RaftMessage(ctx context.Context, query *protos.Payload) (*protos.Payload, error) {
-	if ctx.Err() != nil {
-		return &protos.Payload{}, ctx.Err()
-	}
-
-	for idx := 0; idx < len(query.Data); {
-		x.AssertTruef(len(query.Data[idx:]) >= 4,
-			"Slice left of size: %v. Expected at least 4.", len(query.Data[idx:]))
-
-		sz := int(binary.LittleEndian.Uint32(query.Data[idx : idx+4]))
-		idx += 4
-		msg := raftpb.Message{}
-		if idx+sz > len(query.Data) {
-			return &protos.Payload{}, x.Errorf(
-				"Invalid query. Specified size %v overflows slice [%v,%v)\n",
-				sz, idx, len(query.Data))
-		}
-		if err := msg.Unmarshal(query.Data[idx : idx+sz]); err != nil {
-			x.Check(err)
-		}
-		if msg.Type != raftpb.MsgHeartbeat && msg.Type != raftpb.MsgHeartbeatResp {
-			x.Printf("RECEIVED: %v %v-->%v\n", msg.Type, msg.From, msg.To)
-		}
-		if err := applyMessage(ctx, msg); err != nil {
-			return &protos.Payload{}, err
-		}
-		idx += sz
-	}
-	// fmt.Printf("Got %d messages\n", count)
-	return &protos.Payload{}, nil
-}
-
-func (w *grpcWorker) JoinCluster(ctx context.Context, rc *protos.RaftContext) (*protos.Payload, error) {
-	if ctx.Err() != nil {
-		return &protos.Payload{}, ctx.Err()
-	}
-
-	// Best effor reject
-	if _, found := groups().Server(rc.Id, rc.Group); found || rc.Id == Config.RaftId {
-		return &protos.Payload{}, x.Errorf(errorNodeIDExists)
-	}
-
-	node := groups().Node(rc.Group)
-	if node == nil {
-		return &protos.Payload{}, nil
-	}
-	node.Connect(rc.Id, rc.Addr)
-
-	c := make(chan error, 1)
-	go func() { c <- node.AddToCluster(ctx, rc.Id) }()
-
-	select {
-	case <-ctx.Done():
-		return &protos.Payload{}, ctx.Err()
-	case err := <-c:
-		return &protos.Payload{}, err
-	}
 }
