+++ worker/export.go
@@ -340,7 +340,7 @@ func export(gid uint32, bdir string) error {
 func handleExportForGroup(ctx context.Context, reqId uint64, gid uint32) *protos.ExportPayload {
 	n := groups().Node(gid)
 	if n != nil && n.AmLeader() {
-		lastIndex, _ := n.store.LastIndex()
+		lastIndex, _ := n.Store.LastIndex()
 		n.syncAllMarks(n.ctx, lastIndex)
 		if tr, ok := trace.FromContext(ctx); ok {
 			tr.LazyPrintf("Leader of group: %d. Running export.", gid)
